requirements.txt
eventlet==0.24.1
Flask==1.0.2
Flask_SocketIO==2.8.6
gunicorn==19.7.1
pyserial==3.4
